requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiohttp-client-cache==0.6.1
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==21.4.0
Brotli==1.0.9
cattrs==22.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
exceptiongroup==1.0.4
frozenlist==1.3.3
homeassistant-api==4.0.1
idna==3.4
inky==1.4.0
itsdangerous==2.1.2
multidict==6.0.3
numpy==1.23.5
Pillow==9.3.0
pycares==4.2.2
pycparser==2.21
pydantic==1.9.0
python-forge==18.6.0
PyYAML==6.0
requests==2.28.1
requests-cache==0.9.7
RPi.GPIO==0.7.1
simplejson==3.18.0
six==1.16.0
smbus2==0.4.2
spidev==3.6
typing_extensions==4.4.0
url-normalize==1.4.3
urllib3==1.26.13
yarl==1.8.2
